feat: Add todo creation endpoint
This commit adds a new endpoint for creating todos. The `/todos` route now accepts a POST request with a JSON payload containing the `title` field. Upon receiving the request, the server generates a unique ID for the todo, sets the status to "active", and records the current timestamp as the creation time. The newly created todo is then added to the list of todos.